requirements.txt
grpcio==1.54.2
pymssql==2.2.7
python-decouple==3.8
